TAG v1.6.0.7 main changes: 25501 <platform> <fix> shell task can not sleep after wakeup when 2s timeout 25443 <security> <perf> PMP entry optimization 25641 <pm> <fix> fix vote overflow problem 25760 <platform> <feat> add desc 25827 <feat> synchronize the modifications of lvgl on the 3.1inch branch to the security branch 25691 <wifi> <perf> cali saradc by FT 25148 <wifi> <perf> solve the problem of BLE not broadcasting 25009 <wifi> <perf> optimize saradc 25899 <wifi> <fix> Reschedule channel when next_tbtt is past in chan_tbtt_updated 25688 <platform> <fix> The number offset of GPIO needs to be 64bits 26030 <perf>disable the peripheral interrupt before enter deepsleep, in void to affact the deepsleep 25911 <wifi> <fix> change MAC_RATES_MAX_LEN to 12 25912 <wifi> <fix> check whether the rate is valid when fast connecting 26182 <platform> <test> improve flash_test cmd 26236 <platform> <fix> fix links 26068 <platform> bk libs recreate,change manifest 26449 <platform> <feat> Solve hyperlink jump problem 26839 <wifi> <fix> set MAX_BSS_LIST to 64 26842 <wifi> <fix> beacon channel and operation channel are different,and beacon sequence number restart,maybe ap reopen but not receive deauth,need reconnect 26846 <wifi> <perf> optimize SRRC 27084 <platform> <feat> added mailbox channel reset API 27088 <platform> <feat> amp lock reset after slave cpu is reset 26483 <wifi> <fix> Incomplete synchronization of release branch code leads to bam assert 27091 <wifi> <fix> fix ASSERT at rxl_reset:1094 27089 <wifi> <fix> 1.when auth resend, may disconnect; 2.when reconnect deauth send timeout, it will disconnect. then deauth frame cfm come back with success, it will resend auth. then sm state will incorrect 27090 <perf> hal_machw_reset trace optimize 27226 <wifi> <log> fix ASSERT at rxl_rxdesc_ready_for_processing:1144. 27169 <wifi> <fix>Increased RF interrupt protection range 27092 <wifi> <log> Add some debug information to analyze assert issue in RX process 27570 <wifi> <fix> Fix assert in RX AMSDU process 28176 <platform> <feat> reset mailbox channel when closed 23334 <platform> <fix> time compensate in 1s timer 27880 <platform> <perf> change __FILE__ macro to relpath 22899 <feat> release version config dump feature 23337 <platform> <fix> memdebug check last 4 bytes tag for param malloc size may not 4 bytes align 23338 <platform> <fix> cpu1 pmp protect solution 26487 <platform> <feat> flush i-cache 25437 <platform> <fix> reset reason for software reboot sometimes change to interrupt watchdog 28187 <fix> Fix unexpected assert caused by incorrect AC number 28191 <perf>Frame optimization for lvgl when vertical lcd used as horizontal lcd. 28190 <feat>lvgl lib support use psram as canvas buffer. 28457 <wifi> <perf> Adjust the value range of rssi_offset 28482 <wifi> <fix> Improve the stability of connection with HE AP 28477 <wifi> <perf> Fix Temperature compensation adjustment power causes overflow problem 29377 <platform> <fix> B Area system(or CPU1) crashed. 29369 <platform> <fix> cpu1 pmp protect solution 29385 <fix>resolve psram leak when lcd need to anticlock 90 to match ui. 29425 <perf>add config to support tp's coordinate is not same as lcd. 29596 <wifi> <fix> fix tcp connect assert 26210 <wifi> <perf> optimize txdc in high and low temp 29912 <wifi> <fix> 1、the MAC_HTCAPA_TX_STBC and MAC_HTCAPA_RX_STBC_MSK flag is not checked when beacon is received；2、MAC_HTCAPA_40_MHZ changes not disconnect 29740 <wifi> <log> mm trace optimize 30193 <wifi> <fix> fix ping fail 30078 <wifi> <fix> fix softap cant send udp broadcast pkt 30202 <fix> Add null pointer detection for cameras without UAC 30339 <wifi> <fix> fix ping fail PartII 30351 <wifi> <fix> update mon_ie_crc and mon_ie_crc_cnt after the CSA complete 30719 <wifi> <perf> optimize EPA RF 30871 <wifi> <fix> Save/Get cali_results to/from BK_PARTITION_NET_PARAM 31021 <fix> check MAC_ELTID_HT_CAPA changes before checking valid 30945 <wifi> <fix> Dump happend in stability_WIFI_STA_AP_start_stop_8H_004 31322 <wifi> <perf> optimize saradc in mutli-CPU 38403 <platform> <perf> optimize adc cmd 38433 <wifi> <perf> optimize vnd_cal.exe,add version log

+++ middleware/driver/gpio/gpio_driver_base.c
@@ -531,7 +531,7 @@ static void gpio_config_wakeup_function(void)
 		gpio_wakeup_set_pin_voltage_status(gpio_wakeup_map[i].id, gpio_wakeup_map[i].int_type);
 		bk_gpio_set_interrupt_type(gpio_wakeup_map[i].id, gpio_wakeup_map[i].int_type);
 		//bk_gpio_enable_interrupt(gpio_wakeup_map[i].id);
-		wakeup_id |= 0x1<<gpio_wakeup_map[i].id;
+		wakeup_id |= (uint64_t)0x1<<gpio_wakeup_map[i].id;
 		s_gpio_is_setted_wake_status |= ((uint64_t)1 << gpio_wakeup_map[i].id);
 	}
 
@@ -555,7 +555,7 @@ static void gpio_config_wakeup_function(void)
 			gpio_wakeup_set_pin_voltage_status(s_gpio_dynamic_wakeup_source_map[i].id, s_gpio_dynamic_wakeup_source_map[i].int_type);
 			bk_gpio_set_interrupt_type(s_gpio_dynamic_wakeup_source_map[i].id, s_gpio_dynamic_wakeup_source_map[i].int_type);
 			//bk_gpio_enable_interrupt(s_gpio_dynamic_wakeup_source_map[i].id);
-			wakeup_id |= 0x1<<s_gpio_dynamic_wakeup_source_map[i].id;
+			wakeup_id |= (uint64_t)0x1<<s_gpio_dynamic_wakeup_source_map[i].id;
 			s_gpio_is_setted_wake_status |= ((uint64_t)1 << s_gpio_dynamic_wakeup_source_map[i].id);
 		}
 	}
